[ui] update backfill cancel button to only have one option (#27272)
## Summary & Motivation
Previously the backfill cancellation button had two titles "Cancel
backfill submission" and "Terminate unfinished runs". Over time it seems
that most of the cases when "Terminate unfinished runs" was shown got
phased out, and "Cancel backfill submission" is a misnomer because it
doesn't stop the backfill from being added to the DB, it cancels the
backfill and all runs it has launched.

Additionally, in the remaining case when "Terminate unfinished runs" was
shown, it used a different code path to just cancel the runs launched by
the backfill and didn't cancel the backfill itself, potentially leading
to a confusing state where a user has "canceled" the backfill but the
backfill is still running and submitting new runs.

This PR consolidates the button so that it just has one behavior:
sending a cancellation query to the backfill. This will signal to the
daemon to cancel any in-progress runs, then mark the backfill as
canceled. Additionally it allows us to immediately exit the dialog that
appears when a user cancels a backfill. The button is also renamed to
make its action more clear

Screen recording of canceling a backfill

https://github.com/user-attachments/assets/345a27b9-aa07-4aeb-9960-c6524251565e

If a user doens't have permission to cancel a backfill or the backfill
cannot be canceled (already completed), the button is greyed out
<img width="1495" alt="Screenshot 2025-01-22 at 12 02 39 PM"
src="https://github.com/user-attachments/assets/92219355-1264-47dd-91f4-a7fe48472722"
/>


## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.