[components] Update docs commenting logic (#27162)
## Summary & Motivation

This simplifies the docs commenting logic by re-using the source-position loader. In essence, we dump the yaml, then parse it again so we can map the values to source positions, then emit comments on lines where we need to.

This is a lot easier to build off of and less sketchy than the previous implementation.

## How I Tested These Changes

## Changelog

NOCHANGELOG