[components] Explicitly pass params to Component.load (#27174)
## Summary & Motivation

This makes way for a new PythonComponentDecl that I'd like to add upstack, but I also just think this is net a better implementation regardless.

In particular, the fact that we needed to call `loaded_params = context.load_params(cls.get_schema())` in every single `load()` method was a big red flag.

## How I Tested These Changes

## Changelog

NOCHANGELOG