Allow concurrency limits to handle changing the default limit (#27112)
## Summary & Motivation
This PR adds a default flag to the storage layer, to help track which
limits inherit from the default versus limits that are explicitly set.
This allows us to change the default and helps us to change those limit
values.

Currently, the `concurrency > default_op_concurrency_limit` setting is
only used to initialize the number of slots for "unconfigured" keys.
Now, we keep track of the slots that are initialized from a default
value and update them if the default value has changed.

## How I Tested These Changes
BK

## Changelog
- Adds the ability to distinguish between explicitly set pool limits and
default-set pool limits. Requires a schema migration using `dagster
instance migrate`.