separate configured limit from slot count (#26803)
## Summary & Motivation
Currently, the event log storage returns a `ConcurrencyKeyInfo` to
represent the information about a particular pool. This info includes
the number of slots available, but it doesn't have enough information to
distinguish between the following cases:

1. no pool limit is explicitly set and there is no default value
2. no pool limit is explicitly set but there is a default value
3. a pool limit is explicitly set to 0

This PR adds more information to the returned object to help distinguish
between cases 1 and 2. Specifically, there's a new nullable `limit`
value, which returns the configured limit regardless of whether the pool
has been initialized to create the corresponding slot rows. A limit
value of `None` indicates an unconfigured pool, with no default limit.

## How I Tested These Changes
BK