[ui] Fix launchpad warning about missing partition for hidden asset job
This is a follow-up to https://github.com/dagster-io/dagster/pull/25971/files

Resolves https://linear.app/dagster-labs/issue/FE-765/non-partitioned-run-of-partitioned-job

In a release this fall, we combined all the hidden asset jobs into a single `__ASSET_JOB` that spans all assets. The objective of this was to enable partition-mapping and allow users to materialize any arbitrary subset of assets.  To do this, the partition sets were removed from the hidden job and the appropriate partition set is determined on the fly based on the assets within the job you’re interested in.

This broke the warning in this modal, since the job has partitioned assets but does not have `partitionSets`.

This looks like a big code change but I just took all the existing loader logic out of ConfigEditorConfigPicker’s partition set + partition picker, and moved it up one layer so that it could be referenced in the warning as well.