store end timestamp on PartitionBackfill (#27283)
## Summary & Motivation
We don't store the time a backfill ends on the `PartitionBackfill`
object. Instead we query for all of the runs in the backfill and find
the one with the most recent end time. This results in noticeably slow
load times (~10seconds on our deployment) on the Backfills tab of the
Runs page because we have to fetch every run for every backfill on the
page.

The PR stores the current time when a backfill is moved into a terminal
state. Then we can access this value in the GQL resolver. If the end
timestamp doesn't exist on the backfill object we fall back to
calculating based on run end times. Maybe we could also write this value
to the db after we've computed it?

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.