[ui] Fix long asset key display in header (#27349)
## Summary & Motivation

We received a report that long asset key path values are broken in the Asset page header breadcrumbs. We had fixed this before, but a recent change to fix some incorrect React node re-use looks like it also inadvertently affected measurement for truncation.

To fix this, remove wrapper spans and put the key on the `TruncatedHeading` elements themselves.

I'm not sure if this will reintroduce the flashing issue noted in https://github.com/dagster-io/dagster/pull/25188, but I don't see any flashing when I test this with very long keys.

## How I Tested These Changes

View an asset with long keys, verify truncation.

## Changelog

[ui] Fix Asset page header display of long key values.