[old docs] Fix uncontrolled path expression, run lint (#27363)
## Summary & Motivation

Aims to resolve
https://github.com/dagster-io/dagster/security/code-scanning/26.

Sanitize the file path being sought for markdown to ensure that the
requested file actually lives inside the target `content` directory.

## How I Tested These Changes

`yarn dev`, verify that navigation still works. Try to navigate to a
bogus path, verify that an error is correctly thrown. (This behavior is
effectively unchanged and the user would have arrived at a 404 anyway.
This was just a sanity check.)