[components] Ensure dg component check uses configured components path and other tweaks (#27386)
## Summary & Motivation

This makes a few tweaks around the new `dg component check` command and
it's tests:

- Adds a clause to ensure this is running inside a code location
(something we do for all `dg component ...` commands)
- Change to use the parametrized component path derived from context
instead of a hardcoded components path (which will break if the user
tries to configure the components path)
- Does a little refactoring around the tests to make them more coherent
with the rest of `dagster-dg` tests. In particular:
- makes the `create_code_location_from_components` utility build off of
`isolated_example_code_location_foo_bar`, which is used everywhere else
    - nests the `check` tests with the other `dg component ...` tests

## How I Tested These Changes

New unit tests.