poll every n run submissions to see if sensor is off (#26958)
## Summary & Motivation
While submitting runs, polls the status of the sensor after every N runs
are submitted. If the sensor gets turned off, we stop submitting runs
and mark the tick as SKIPPED so that we don't try to resubmit the
remaining runs in the future.

The motivation for this is to allow users to stop sensors if they are
submitting a large number of unexpected runs (usually due to a
configuration error or some other error in the sensor logic). They'll
want to 1. prevent the remaining runs from getting submitted, and 2. fix
the sensor before starting it back up. When they do start it back up, we
want to make sure we don't continue submitting the runs from the old/bad
tick, which is why we mark the tick as skipped.

internal pr for the same setting
https://github.com/dagster-io/internal/pull/13312

To do:
- [x] name the setting
- [x] figure out if the different behavior seeing in the test is a bug
or expected
- [x] make sure we are fine not reverting the cursor 
- [x] UI - having internal design discussion

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.