data migration for backfill end timestamp (#27325)
## Summary & Motivation
Data migration for adding an end timestamp for backfills. 

## How I Tested These Changes
backcompat unit tests. Also ran the migration on my local db, and 1.
visually confirmed that the end timestamp was added for backfills; 2.
loaded the ui page that accesses the end timestamp and confirmed it did
not call the path that queries all runs for a backfil.

## Changelog

Optional data migration to improve performance of the Runs page. Run
`dagster instance migrate` to run the date migration. The migration will
update serialized backfill objects in the database with an end timestamp
attribute computed by querying the runs launched by that backfill to
determine the when the last run completed.