[ui] Fix launchpad warning about missing partition for hidden asset job (#27340)
This is a follow-up to
https://github.com/dagster-io/dagster/pull/25971/files

Resolves
https://linear.app/dagster-labs/issue/FE-765/non-partitioned-run-of-partitioned-job

## Summary & Motivation

In a release this fall, we combined all the hidden asset jobs into a
single `__ASSET_JOB` that spans all assets. The objective of this was to
enable automatic-partition-mapping and allow users to materialize any
arbitrary subset of assets. To do this, the partition sets were removed
from the hidden job and the appropriate partition set is determined on
the fly based on the assets within the job you’re interested in.

This broke the warning in this modal, since the job has partitioned
assets but does not have `partitionSets`.

This looks like a big code change but I just took all the existing
loader logic out of ConfigEditorConfigPicker’s partition set + partition
picker, and moved it up one layer so that it could be referenced in the
warning as well. The queries / transformations of the data, etc. haven't
been changed.

## How I Tested These Changes

This was fairly easy to repro once I figured it out - just go to the
graph (not an asset job) and choose "Open Launchpad" from the
materialize dropdown with a partitioned asset selected.

## Changelog

[ui] When launching partitioned assets in the launchpad from the global
graph, Dagster will warn if you have not made a partition selection.

Co-authored-by: bengotow <bgotow@elementl.com>