[components] Make dg component-type list output a table with colors (BUILD-637) (#27381)
## Summary & Motivation

Linear: https://linear.app/dagster-labs/issue/BUILD-637/generate-a-table-output-for-dg-component-type-list

Changes the output of `dg component-type list` to be a table. Uses the `rich` library (which is already a dependency of `typer`). Probably we will eventually completely remove the `typer` dependency and just use `rich` directly to style all output.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/YRZmeX8NtwrdWNuUQ4Fv/1c257d58-48ff-4a40-b3e7-72c825663a49.png)

## How I Tested These Changes

Adjusted existing tests.