[components] Provide good error message if dagster-components is not available (BUILD-633) (#27388)
## Summary & Motivation

Currently some `dagster-dg` commands fail with an ugly error message if `dagster-components` is not found on the path. This can happen if you are not using a dg-managed environment.

This PR runs a check for `dagster-components` on the path (when it is needed) and also does some refactoring of our validation code more generally.

Prior error message was an uncaught subprocess error (`dagster-components` not found) with a backtrace.

New error message:

```
Could not find the `dagster-components` executable on the system path.

The `dagster-components` executable is installed with the `dagster-components` PyPI package and is necessary for `dg` to interface with Python environments containing Dagster definitions. `dagster-components` is installed by default when a code location is scaffolded by `dg`. However, if you are using `dg` in a non-managed environment (either outside of a code location or using the `--no-use-dg-managed-environment` flag), you need to independently ensure `dagster-components` is installed.
```

## How I Tested These Changes

New unit tests.