[components] Remove CodeLocationProjectContext from dagster-components (#27391)
## Summary & Motivation

When `dagster-components` split from `dagster-dg`, it left behind a duplicated `CodeLocationContext` object that has been a source of pain and confusion ever since. This was used in the CLI but never in a consistent way-- it also does some half-baked reading of `pyproject.toml` to read settigs that are also read in `dagster-dg`.

This PR completely cuts out `CodeLocationProjectContext` in `dagster-components`, turning the entire idea of "project context" into a `dg`-only thing. When `dagster-components` needs some information derived from `dg`'s config, it is expected to receive this via the CLI call.

The main impact on external APIs here is that a few of them have been changed to expect the component instances root directory as argument rather than a code location root directory-- `dagster-components` no longer knows anything about project structure, it just knows how to make a directory and traverse it for component instances.

## How I Tested These Changes

Modified existing test suite.