[components] Better-scoped test environments (#27392)
## Summary & Motivation

Prior to this PR, we were using generated code locations, which scaffold a new venv containing `dagster-components`, for the environment for most tests. When we wanted to test in a non-code-location env we did some hackery.

This PR adds fixtures to generate a generic virtual environment containing `dagster-components` as well as a component library (but non-code-location project). These are more appropriate fixtures to use for commands that don't rely on a code location.

## How I Tested These Changes

Modified existing unit tests.