[components] Rework resolvable model (#27442)
## Summary & Motivation

This is a larger rework of how the yaml resolution logic works in service of creating a clear separation between the `__init__` args and the params schema.

Each params schema is now a `ResolvableModel`, which defines an explicit function to turn the blob of configuration into an explicit object type.

This also removes the "first pass" of the resolution logic, which was way too confusing.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.