Keep sys.path altered after the initial import of user code
Summary:
The goal of cleaning after ourselves here is understandable, but prevents user code that depends on being able to re-import the same module implicitly (like ProcessPoolExecutor) from working correctly.

Fixed an issue where assets using ProcessPoolExecutor or multiprocessing with a spawn context sometimes failed to work with a ModuleNotFoundError.

> Insert changelog entry or delete this section.