[components] Standardize terminal width in dagster-dg help output tests (BUILD-646) (#27484)
## Summary & Motivation

The width of the help output generated by certain tests in `dagster-dg`
varied depending on environment. This caused these tests to fail when
run locally with `tox` (but oddly not my base env). This PR modifies the
tests to ensure a fixed width in all environments.

## How I Tested These Changes

Ran existing tests in tox, they now pass.