[Selection Input] Refactor auto-complete system. (#27389)
## Summary & Motivation

The primary goal of this refactor is to decouple the auto-complete
suggestion UI from the autocomplete visitor logic.

Before:
- Visitor hard coded the types of the suggestions 
- UI rendering was tightly coupled to those types
- The types were too restrictive, they required all of the values to be
strings but that doesn't work well for tags which have a key and a
value. Encoding the tags as a string (eg: `"key"="value"`) created
ambiguities, where we could not tell if the string was a tag of `{key:
"key", value: "value"}` or a tag of `{key: "key"="value", value:
undefined}`.


After:
- Rendering is encapsulated in the SelectionAutoCompleteProvider which
is also responsible for making auto-complete suggestions and can handle
both string and tag object types.



## How I Tested These Changes

Existing jest test in SelectionAutocomplete.ts. + Manual testing in
cloud and OSS.