[cli-refactor] Standardize format and naming of click option decorators (#27492)
## Summary & Motivation

This is the base of a stack that aims to generally improve up the
`dagster` CLI. The goals of the stack are to:

- Establish strong typing by immediately validating untyped bags of
"kwargs" and converting to typed value objects at all entry points
(currently we pass these bags of "kwargs" fairly deep into our code)
- Ensure consistency across CLI entry points. There are many shared
options, we want to make sure validation is being done the same way,
etc.
- Reduce code volume and improve clarity. The current code is quite
convoluted and often inconsistent in its terminology, both internally
and with abstractions deeper in the codebase.

This first PR:

- Updates the name of shared click option groups to use "options"
instead of "argument", e.g. `python_origin_target_argument` ->
`python_origin_target_options`.
- Standardizes naming across functions that create `click` options and
those that can actually be used as decorators. Currently the naming is
confusing, making it unclear what can be used as a decorator.

## How I Tested These Changes

Existing test suite.