new grpc proto file name for dagster (#27498)
## Summary & Motivation
This is called 'api.proto' which is a very generic name - if another
grpc project is imported with that name, they can't coexist. Rename ours
to 'dagster_api.proto' instead (this is not a breaking change from
grpc's perspective).

## How I Tested These Changes
BK, sanity check that the change isn't breaking by launchign a grpc
server in master and connecting to it with dagster dev with this change
applied, and vice versa

## Changelog
Fixed an issue where dagster could not be imported alongside some other
libraries using gRPC with an 'api.proto' file.