[docs-beta][tool] Fix links to /index pages from docs beta deploy GH Action (#27513)
## Summary

The docs beta deploy GitHub Action constructs links to changed pages
based on the filepath, stripping `.mdx`. This doesn't work for index
pages, which do not include `index` in the URL.

Strips out a trailing `index` before the `.md(x)` suffix on a file, when
turning it into a URL for the GitHub Action comment.

## Test Plan

See below.