[dagster cli] Add option to dagster definitions validate to load in-process (#27459)
## Summary

By default, `dagster definitions validate` starts up a gRPC server to
load the code location. This adds a decent amount of runtime to the
command (locally, for me, about 1s of a total 2.6s runtime).

This PR instead defaults to loading the code location in-process, unless
pointing at multiple code locations (via workspace.yaml) or a location
which specifies an `executable_path`. This makes the command pretty
quick, with the runtime almost entirely dominated by importing the
user's code and loading the defintions (around 1.6s for a full
components project for me).

## How I Tested These Changes

Update unit tests.

## Changelog

> By default, `dagster definitions validate` command now loads locations
in-process, which speeds up runtime.