[components] Update component naming scheme (#27512)
## Summary & Motivation

This updates the naming scheme of our components to be `<name>@<namespace>` instead of `<namespace>.<name>`.

This also updates some terminology throughout the codepaths to distinguish between the name of a component (which does not include the namespace), and the registry key (which is the unique identifier). Previously, we used the term "name" for both of these cases, which was confusing.

## How I Tested These Changes

## Changelog

NOCHANGELOG