[components] Synchronize dagster-dg and dagster-components ComponentKey class (#27539)
## Summary & Motivation

This separates out the ComponentKey class into its own separate file in both dagster-dg and dagster-components, and updated dagster-components to distinguish between local and global keys.

Unfortunately, we're forced to copy-past code for the time-being as there's not a good way of sharing code between these libraries, but at least now it's easy to copy paste between the files to keep them in sync.

## How I Tested These Changes

## Changelog

NOCHANGELOG