[dagster-dbt] Fix error when using dbt_assets decorator with io_manager_key (#27548)
## Summary & Motivation

Fixes a regression in dagster-dbt after converting it to use spec-based
creation. The core thing here is that we need a way to pass a
DagsterType into the system when using specs.

Before, it was assumed that if you were passing an io_manager_key, this
meant that you definitely wanted that IOManager to handle all of the
outputs, hence the DagsterType was set to DagsterAny instead of the
default Nothing.

However, this is not actually a safe assumption in all cases, as with
the dagster-dbt integration it's useful to set the io_manager_key so
that downstream assets can know how to consume these outputs.

This PR follows patterns we've used in the past of adding another bit of
named metadata to the spec that allows us to wire this sort of
information into the op construction logic.

Because this is a fairly niche usecase, I did not add a
`with_dagster_type` method to the AssetSpec (which would further copy
the io_manager_key solution), but that's something we could consider in
the future.

## How I Tested These Changes

Unit tests

## Changelog

Fixed a bug introduced in dagster-dbt 0.25.7 which would cause execution
to fail when using the `@dbt_assets` decorator with an `io_manager_key`
specified.