Add basic utility/class for WebWorkers (#27576)
## Summary & Motivation

1. Added a `createWorkerThread` utility to be used inside of web-worker
threads. This utility handles setting up feature flags and forwarding
errors back to the main thread. In cloud it will also receive cloud
specific flags.
2. Added a `Worker` class to be used in place of the global Worker
class. This class forwards feature flags and receives errors. It's a
shared component so that cloud can inject custom behavior.


In a follow-up PR I will figure out mocking with these utilities and add
lint rules to make us go through our custom Worker abstraction once we
feel its ready for prime time.

## How I Tested These Changes

I stopped forwarding the `flagSelectionSyntax` flag manually for
`ComputeGraphData.worker` and saw that it was correctly handling the
selection syntax when the flag was on and not when it was off.