[components] Fix SlingResource resolution (#27643)
## Summary & Motivation

After recent changes, we no longer do a pre-resolution step for the params schema, which means we need to explicitly resolve parameters passed to resources.

This should probably be done using a separate wrapper class but for now we can just do the simple fix that restores previous behavior.

A similar thing should be done for dbt

## How I Tested These Changes

## Changelog

NOCHANGELOG