[ui] Repair asset check asOf param from run logs (#27628)
## Summary & Motivation

Asset check events in the run logs link to a seconds-based timestamp on the corresponding asset permalink, since the value coming from GraphQL is seconds-based. The value on the permalink query param needs to be msec-based.

I think it would be better if all of our timestamp fields consistently provided msec-based timestamps, but even within the relevant GraphQL types involved in this narrow use case (asset evaluation and materialization timestamps), I see a mix of seconds-based and msec-based...so this is potentially a gnarly change to take on here. cc @OwenKephart 

## How I Tested These Changes

View run logs where the issue repros. Filter to `type:ASSET_CHECK_EVALUATION`, verify that permalinks to assets have correct `asOf` values, and that navigation to those pages results in appropriate timestamps in the UI.

## Changelog

[ui] Fix timestamped links to asset pages from asset check evaluations in run logs.