Ignore vale failures (#27666)
My understanding of this doc step is that we eventually want to use it,
but we consider it non-blocking on docs PRs today?

In that case, should we tell GitHub Actions that it's okay for this step
to fail? I think that'll prevent all of our docs PRs from having a big
red failure status. Thoughts?