[ui] Run timeline: Fix group-by persistence (#27623)
## Summary & Motivation

Storage persistence of the group-by option on the run timeline is currently broken. More specifically, the value may be persisted, but it's not being used -- the default value (`job`) always ends up as the initial value.

I'm not sure that the default behavior of `useQueryAndLocalStoragePersistedState` is quite right for this use case (namely, `Syncs changes back to localStorage`) so I'm trying to simplify the behavior here specifically without changing `useQueryAndLocalStoragePersistedState` itself and risking breaking other callsites.

- If the query param is present, use it.
- If not, use the stored value.
- If that's not set, use the default value.
- When changing the group-by option with the dropdown, update the query param and set the localStorage value.

Ideally this should allow sharing the URL with the query param applied without affecting the user's stored value, while also allowing the stored value to be used when the query param is absent.

## How I Tested These Changes

View run timeline with no query param.

- Verify that my stored group-by value is applied. ("automation")
- Change group-by value to "job". Verify that storage is updated and that the query param `groupBy=job` now appears in the URL.
- Reload the page with query param `groupBy=automation`. Verify that the query param is used.
- Remove the query param, load the page. Verify that the page reverts to "job", since that is my stored value.

## Changelog

[ui] FIx persistence of group-by setting in run timeline view.