[Asset Selection] Add column/column_tag/table_name/changed_in_branch filters to the grammar. (#27652)
## Summary & Motivation

We want to allow these filters to be represented in the backend (eg. for
catalog views) without actually being able to resolve them.

This means on the frontend we'll need to add custom logic to highlight
that some of these filters are unsupported outside of branch deployments
and dagster plus. (eg. column_tag/table_name/column are only available
if elastic search is available and changed_in_branch is only available
in branch deployments).

## How I Tested These Changes

Existing jest tests + pytest