[core-api][experimental] include_sources -> include_external_assets (#26710)
## Summary & Motivation

decision: experimental -> GA (after parameter change)
reason: if we're going to make this change, we should acompany it with
marking this API as stable (i.e. we should only make the name change
once). "source assets" as a concept no longer really exist, and external
assets is the more general term (and more accurate to what these
parameters were actually filtering for.
docs exist: yes (api only, which is ok)

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.