[core-api] base (#26770)
## Summary & Motivation

This is the base of the core-api-audit stack. The entire stack will be merged in at the 1.10 release, and for now to keep things tidy, changes will be merged downwards as they are approved.

This means that this PR contains only code that has been reviewed and approved in a different PR.

## How I Tested These Changes

## Changelog

The `include_sources` param on all `AssetSelection` APIs has been renamed to `include_external_assets`.