Make asset backfill eligibility logic subset-aware instead of iterating through partition keys individually (#25997)
## Summary & Motivation
This PR changes the core asset backfill eligibility logic to make it
work with subsets instead of individual asset partition keys. This makes
things more complicated (instead of returning a boolean + string in the
bfs logic, we have to return the subset that passed and the subsets +
reasons that they didn't pass, but much faster for large partition sets.
A backfill of 125k partition keys that used to take 20 minutes to
compute now takes about 10 seconds.

## How I Tested These Changes
asset back, dry runs on existing large backfills

## Changelog
Performance improvements for backfills of large partition sets.