[components] Enforce that all component schemas are of type ResolvableModel (#27677)
## Summary & Motivation

Doing this has a few benefits. The first is that we force people to use ResolvableModel as the base class for their component schemas, which means we're able to inject additional stuff into that class in the future.

The second is that we can create a default `load()` method, which pretty significantly reduces boilerplate for these component types.

## How I Tested These Changes

## Changelog

NOCHANGELOG