move dagster dev tests to core (#27654)
## Summary & Motivation

Our only `dagster dev` tests currently live in the daemon integration
test suite, which isn't run on windows. This moves them to `dagster`,
which is tested on windows.

The tests are mostly moved over as they were, with a few adjustments:

- Instead of loading the toys repo, we now just load the simple test
repo loaded in the rest of the tests. The toys repo adds a bunch of
dependencies but loading it specifically doesn't really test anything
relevant.
- All tests have been augmented to ensure all child processes are shut
down (on Unix at least-- see below)
- Some refactoring to make the tests simpler to read.

Previously we weren't testing this on Windows at all, and I went in
thinking we would end with exactly the same tests passing on windows and
unix. Unfortunately, that's not _quite_ true-- with our current set up I
could not find a way to reliably and cleanly shut down the `dagster dev`
process tree on Windows in a CI environment. This has to do with the
differences in signal handling and process groups in Windows vs unix. I
will quote ChatGPT-o1:

>On Windows, CTRL_BREAK_EVENT (and similarly CTRL_C_EVENT) only works
when the calling process and the target processes share a console. By
default, Python’s `subprocess.Popen` with CREATE_NEW_PROCESS_GROUP
spawns the child into its own new process group but does not
automatically attach it to the same console as the parent (and in many
CI environments, there may be no real console at all). As a result,
sending CTRL_BREAK_EVENT to that child process group essentially does
nothing—Windows never delivers the interrupt to those child processes,
and they keep running.

The lack of console in the CI environment means that `CTRL_BREAK_EVENT`,
which we use to shut down the child processes, does not propagate
directly from `dev` to the webserver/daemon (but it does appear to work
when `dagster dev` is used interactively, due to sharing a console). So
in the Windows tests I ended up explicitly killing all child processes.
We could maybe make this more robust in the future by having the `dev`
process send a shutdown signal via another form of IPC, but I couldn't
see how to do that here.

## How I Tested These Changes

Ran moved `dagster dev` tests on windows and unix.