Remove fixed-server-id tracking from grpcserverregistry
Summary:
This is causing issues when using the proxy server entrypoint because we set a fixed server ID (the value of which the frontend uses when deciding whether to reload information from the backend) and then don't update it when the underlying code server process changes. Instead, just rely on the code servers themselves to tell us what the underlying server ID is, which will automatically be updated correctly in both entrypoints when the process is reloaded.

Test Plan:
Load definitions in dagster dev with one resource, add a second resuorce, reload definitions in the UI
Before: server ID is fixed and does not update, so the frontend doesn't update either
Now: server ID changes, and frontend updates