Kill repository location state observer (#27777)
## Summary & Motivation

This has been succeeded by the Workspace polling work we did last half. 



https://dagsterlabs.slack.com/archives/C03A0D72A6T/p1739310233122249?thread_ts=1739303453.899019&cid=C03A0D72A6T


## How I Tested These Changes
Used app-proxy to load a few customers in cloud and things worked fine.
Not super sure how to test all of the various cases in OSS but going to
rely on Daniel/Alex's word that the Workspace polling should work for
all OSS use cases now.