[record] Fix handling of record classes with pydantic (#27790)
## Summary & Motivation

Prior to these changes, both added tests would fail. Updated AssetSpec to be an `@record` so that it could take advantage of these changes.

The core issue is that the pydantic internals assume that all namedtuple subclasses can be instnatiated with positional args based on their fields, which is not the case for records, which are kw-only (and is not the case for explicit NamedTuple subclasses which define `__new__`).

This was the least-gross way I could find to handle this behavior, and in reality is not particularly invasive to the `record` framework (import is deferred, just a single extra attribute)

## How I Tested These Changes

## Changelog

NOCHANGELOG