[ui] Switch graph traversal operators to BFS to ensure shortest path is used (#27795)
## Summary & Motivation

Now here's a long-standing one! The traversal code used to navigate the
asset graph for `asset++` was doing depth-first and not breadth-first
search, which meant an asset connected via multiple paths of varying
length might not be included in the result. This is important for the
lineage view because under the hood, because a filter of "Distance: 3"
might not include an asset that is connected via both distance=4 and
distance=3 paths


https://linear.app/dagster-labs/issue/FE-779/lineage-ui-bug-displaying-upstream-assets-assigned-to-ui-team

## How I Tested These Changes

I added new tests and also manually verified the specific case in the
ticket is resolved.

## Changelog

[ui] The traversal operators on the asset graph (`asset++`) correctly
include assets connected to the target asset by paths of varying
distance.

---------

Co-authored-by: bengotow <bgotow@elementl.com>