turn on the op concurrency counter for the run queue by default (#27824)
## Summary & Motivation
Turns on the concurrency run-blocking by default... previously, since we
didn't want to change behavior in a minor release, we used configuring a
`pool` setting as a proxy for it. That's no longer needed since this is
for a major release and we should just change the behavior.

## How I Tested These Changes
BK

## Changelog
* Turns on run-blocking for concurrency keys / pools by default. For op
granularity, runs are dequeued if there exists at least one op that can
execute once the run has started. For run granularity, runs are dequeued
if all pools have available slots.